% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{Combine_marker_umap}
\alias{Combine_marker_umap}
\title{UmapPlot based on results from Markers Greedy Search}
\usage{
Combine_marker_umap(scrna, df.split, id)
}
\arguments{
\item{scrna}{seurat objects}

\item{df.split}{results from Greedy Search}

\item{id}{cell groups to identify}
}
\value{

}
\description{
UmapPlot based on results from Markers Greedy Search
}
