% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{Detect_single_marker}
\alias{Detect_single_marker}
\title{Detect single markers of specific cell groups}
\usage{
Detect_single_marker(
  scrna,
  id,
  step = 0.1,
  slot = "data",
  category = NULL,
  geneset = NULL,
  assay = "RNA",
  do.fast = F,
  min.pct = 0.15,
  min.fc = 0.25,
  use.all = F,
  do.f1score = F,
  pseudo.count = 0.01,
  min.tnr = 0.65,
  org = "human",
  self.db = NULL,
  self.db.only = F,
  ...
)
}
\arguments{
\item{scrna}{seurat obj to be used}

\item{id}{interested cell group}

\item{step}{quantile steps}

\item{slot}{indicate the slot of data to compute}

\item{category}{Denotes which experiment to use. Available options are:
\itemize{
\item{"IHC"} : use genes with valid antibody for IHC in human protein atlas
\item{"ICC"} : use genes with valid antibody for ICC in human protein atlas
\item{"ICC.IHC"} : use genes with valid antibody for ICC or IHC in human protein atlas
\item{"Flow"} : Only use cell surface genes with at least one valid antibody(ICC or IHC)
}}

\item{geneset}{custom genes to test}

\item{assay}{indicate the assay to compute}

\item{min.pct}{only test genes that are detected in a minimum fraction of cells in interested cell types. Default is 0.15}

\item{use.all}{Don't do any filter on input geneset}

\item{self.db}{self defined antibody database}

\item{self.db.only}{Bolean. Only use the self deifned antibody database or not}
}
\value{
list of markers performance
}
\description{
Detect single markers of specific cell groups
}
