% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{generate_report}
\alias{generate_report}
\title{generate  html report for single markers of all cell groups}
\usage{
generate_report(
  scrna,
  markers.list,
  fpath = ".",
  aggr.other = F,
  fname = NULL,
  top_n_genes = 6,
  ridge_ncol = 3,
  assay = "RNA",
  slot = "data",
  ...
)
}
\arguments{
\item{scrna}{seurat obj to be used}

\item{markers.list}{results of Detect_single_marker_all()}

\item{fpath}{Path to generate the report}

\item{aggr.other}{whether to aggregate other celltypes}

\item{fname}{file name. Default NULL}
}
\value{
list of markers performance
}
\description{
generate  html report for single markers of all cell groups
}
