% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{generate_report_rmd}
\alias{generate_report_rmd}
\title{generate  html report for single markers of all cell groups}
\usage{
generate_report_rmd(
  scrna,
  markers.list,
  aggr.other = F,
  top_n_genes = 6,
  ridge_ncol = 3,
  fname = NULL,
  fpath = ".",
  assay = "RNA",
  slot = "data"
)
}
\arguments{
\item{scrna}{seurat obj to be used}

\item{markers.list}{results from Detect_single_marker_all()}

\item{top_n_genes}{top number of genes for RidgePlot}

\item{fname}{file name. Default NULL}

\item{fpath}{Path to generate report}

\item{sc2marker_ncol}{number of col of Ridgeplot}
}
\value{
list of markers performance
}
\description{
generate  html report for single markers of all cell groups
}
