% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{get_antibody}
\alias{get_antibody}
\title{Get Antibody information table}
\usage{
get_antibody(
  markers.list,
  rm.noab = T,
  org = "human",
  self.db = NULL,
  self.db.only = F,
  return_matrix = F,
  ...
)
}
\arguments{
\item{markers.list}{makers list from Detect single markers function}

\item{rm.noab}{whether to remove genes without antibody information. Default TRUE.}

\item{self.db}{User provided database}

\item{self.db.only}{Bolean, whether to only use the user provided database}
}
\value{
list of markers with antibody information
}
\description{
Get Antibody information table
}
