% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{get_gene_PRAUC_matrix}
\alias{get_gene_PRAUC_matrix}
\title{Compute Precision and Recall matrix for input gene}
\usage{
get_gene_PRAUC_matrix(scrna, gene, id, step = 0.01)
}
\arguments{
\item{scrna}{seurat obj to be used}

\item{gene}{gene to test}

\item{id}{interested cell group}

\item{step}{quantile steps}
}
\value{
PRAUC of input gene
}
\description{
Compute Precision and Recall matrix for input gene
}
