% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{get_gene_PRAUC_pos}
\alias{get_gene_PRAUC_pos}
\title{Compute PRAUC for positive markers}
\usage{
get_gene_PRAUC_pos(scrna, gene, id, step = 0.01)
}
\arguments{
\item{scrna}{seurat obj to be used}

\item{gene}{gene to test}

\item{ge}{interested cell group}
}
\value{
PRAUC of input gene
}
\description{
Compute PRAUC for positive markers
}
