% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{get_gene_score}
\alias{get_gene_score}
\title{Calculate gene specific score in both direction}
\usage{
get_gene_score(exprs.matrix, id, gene, step = 0.01, pseudo.count = 0.01)
}
\arguments{
\item{exprs.matrix}{exprs matrix with ident panel}

\item{id}{interested cell id}

\item{gene}{gene to use}

\item{step}{quantile step}
}
\value{
gene score in both direction
}
\description{
Calculate gene specific score in both direction
}
