% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{get_gene_score_neg.fbeta}
\alias{get_gene_score_neg.fbeta}
\title{Calculate gene positive specific score with FBeta (Hypergate like approach)}
\usage{
get_gene_score_neg.fbeta(
  scrna,
  gene,
  id,
  step = 0.01,
  assay = "RNA",
  slot = "data"
)
}
\arguments{
\item{scrna}{obj to use}

\item{gene}{gene to use}

\item{id}{interested cell id}

\item{step}{quantile step}

\item{assay}{which assay to use in obj, default RNA}

\item{slot}{which slot to use, default data}

\item{panel}{ID panel in metadata of obj}
}
\value{
seurat obj with re-set Idents
}
\description{
Calculate gene FBeta score with negative direction
}
