% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{get_gene_score_neg}
\alias{get_gene_score_neg}
\title{Calculate gene negative specific score}
\usage{
get_gene_score_neg(scrna, gene, id, step = 0.01, assay = "RNA", slot = "data")
}
\arguments{
\item{scrna}{obj to use}

\item{gene}{gene to use}

\item{id}{interested cell id}

\item{step}{quantile step}

\item{assay}{which assay to use in obj, default RNA}

\item{slot}{which slot to use, default data}
}
\value{
seurat obj with re-set Idents
}
\description{
Calculate gene specific score in negative direction
}
