% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{get_original_alpha}
\alias{get_original_alpha}
\title{Calculate gene negative specific score}
\usage{
get_original_alpha(scrna, gene, x.alpha, assay = "RNA", slot = "data")
}
\arguments{
\item{scrna}{obj to use}

\item{gene}{gene to use}

\item{assay}{which assay to use in obj, default RNA}

\item{slot}{which slot to use, default data}
}
\value{
alpha before normalization
}
\description{
Calculate gene specific score in negative direction
}
