% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{get_split_pos_pos}
\alias{get_split_pos_pos}
\title{UmapPlot based on results from Markers Greedy Search}
\usage{
get_split_pos_pos(
  scrna,
  gene1,
  gene2,
  id,
  step = 0.01,
  assay = "RNA",
  slot = "data"
)
}
\arguments{
\item{scrna}{seurat objects}

\item{id}{cell groups to identify}

\item{df.split}{results from Greedy Search}
}
\value{

}
\description{
UmapPlot based on results from Markers Greedy Search
}
