% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{GetGenes}
\alias{GetGenes}
\title{Get cell names which meet cutoff}
\usage{
GetGenes(
  scrna,
  slot = "data",
  category = NULL,
  geneset = NULL,
  assay = "RNA",
  min.pct = 0.1,
  use.all = F,
  species = "Human"
)
}
\arguments{
\item{scrna}{seurat obj to be used}

\item{slot}{indicate the slot of data to compute}

\item{category}{indicate the ICC/IHC/Flow, if NULL use all genes.}

\item{geneset}{custom genes to test, all genes in obj to use if NULL}

\item{assay}{indicate the assay to compute}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expressed. Default is 0.1}

\item{use.all}{Don't do any filter on input geneset}

\item{id}{interested cell group}

\item{step}{quantile steps}
}
\value{
list of markers performance
}
\description{
Get cell names which meet cutoff
}
