% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{identify_single_marker}
\alias{identify_single_marker}
\title{Compute PRAUC for each gene in input geneset and do the ranking}
\usage{
identify_single_marker(scrna, cellgroup, geneset, step = 0.01)
}
\arguments{
\item{scrna}{seurat obj to be used}

\item{cellgroup}{interested cell group}

\item{geneset}{input geneset}

\item{step}{quantile steps}

\item{gene}{gene to test}
}
\value{
PRAUC of input gene
}
\description{
Compute PRAUC for each gene in input geneset and do the ranking
}
