% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{intersect.ignorecase}
\alias{intersect.ignorecase}
\title{Intersect 2 list ignoring the cases and return the elements in list1}
\usage{
intersect.ignorecase(list1, list2)
}
\arguments{
\item{list1}{list 1 of genes}

\item{list2}{list 2 of genes}
}
\value{
Ridge plot
}
\description{
Intersect 2 list ignoring the cases and return the elements in list1
}
