% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{makeid}
\alias{makeid}
\title{make id for plotting}
\usage{
makeid(scrna, id, panel = "seurat_clusters")
}
\arguments{
\item{scrna}{obj to use}

\item{id}{interested cell id}

\item{panel}{ID panel in metadata of obj}
}
\value{
seurat obj with re-set Idents
}
\description{
Reset active.ident of obj for plotting.
}
