% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{marker_stepbystep}
\alias{marker_stepbystep}
\title{Detect greedy step by step filters for cell group identification}
\usage{
marker_stepbystep(scrna, cellgroup, depth = 2, geneset, step = 0.01)
}
\arguments{
\item{scrna}{seurat object}

\item{cellgroup, }{interested cell group}

\item{depth}{how much steps to filter}

\item{geneset}{Gene set to be used}
}
\value{

}
\description{
Detect greedy step by step filters for cell group identification
}
