% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{normalize}
\alias{normalize}
\title{Min-Max Normalization}
\usage{
normalize(x, ...)
}
\arguments{
\item{x}{Gene expression}
}
\value{
Normalized expression data
}
\description{
Normalize scRNA sequencing data.
}
