% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{plot_filter_stepbystep_first2}
\alias{plot_filter_stepbystep_first2}
\title{Detect greedy step by step filters for cell group identification}
\usage{
plot_filter_stepbystep_first2(scrna, df.split, id, step = 0.01)
}
\arguments{
\item{scrna}{seurat object}

\item{cellgroup, }{interested cell group}

\item{depth}{how much steps to filter}

\item{geneset}{Gene set to be used}
}
\value{

}
\description{
Detect greedy step by step filters for cell group identification
}
