% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{Detect_single_marker}
\alias{Detect_single_marker}
\title{Detect single markers of specific cell groups}
\usage{
Detect_single_marker(
  scrna,
  id,
  step = 0.1,
  slot = "data",
  category = NULL,
  geneset = NULL,
  assay = "RNA",
  do.fast = F,
  min.pct = 0.1,
  use.all = F,
  do.f1score = F,
  pseudo.count = 0.01
)
}
\arguments{
\item{scrna}{seurat obj to be used}

\item{id}{interested cell group}

\item{step}{quantile steps}

\item{slot}{indicate the slot of data to compute}

\item{category}{indicate the ICC/IHC/Flow, if NULL use all genes.}

\item{geneset}{custom genes to test, all genes in obj to use if NULL}

\item{assay}{indicate the assay to compute}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expressed. Default is 0.1}

\item{use.all}{Don't do any filter on input geneset}
}
\value{
list of markers performance
}
\description{
Detect single markers of specific cell groups
}
