% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{get_antibody}
\alias{get_antibody}
\title{Get Antibody information table}
\usage{
get_antibody(markers.list)
}
\arguments{
\item{markers.list}{makers list from Detect single markers function}
}
\value{
list of markers performance
}
\description{
Get Antibody information table
}
