% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{get_gene_score}
\alias{get_gene_score}
\title{Calculate gene specific score in both direction}
\usage{
get_gene_score(
  exprs.matrix,
  celltype,
  gene,
  step = 0.01,
  pseudo.count = 0.01,
  min.tnr = 0.65
)
}
\arguments{
\item{exprs.matrix}{exprs matrix with ident panel}

\item{gene}{gene to use}

\item{step}{quantile step}

\item{id}{interested cell id}
}
\value{
gene score in both direction
}
\description{
Calculate gene specific score in both direction
}
