% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_markers.ident.R
\name{makeid}
\alias{makeid}
\title{Reset cell group idents}
\usage{
makeid(scrna, id, panel = "seurat_clusters")
}
\arguments{
\item{scrna}{seurat obj to be used}

\item{id}{interested cell group}

\item{panel}{Ident panel in seurat obj meta.data}
}
\value{
seurat obj with re-set Idents
}
\description{
Reset cell group idents
}
